Make 'idle' event fire exactly once.
Only emit 'idle' when queue length is zero - otherwise we occasionally end up
in the situation where all pending requests finalize simultenously, and our
runningJobs count gets down to zero, but we have a bunch of queued jobs which
will be started on the next tick.  This makes it very difficult to wait for
a bunch of queued jobs to finish.

Emit 'idle' event after calling the callback - that seems like more predictable
behavior.